requirements.txt
Django==1.11.29
django-bootstrap3==12.1.0
djangorestframework==3.11.2
Pillow==8.0.1
pkg-resources==0.0.0
pytz==2020.4
